Updated/Fixed the send error handling
The send error handling was wrapped in a try/catch that ignored any errors. Further
the fwrite operation had an @ prepended to it which would suppress any errors before
they could happen. fwrite() does not throw any exceptions, it works off return values
to determine success. Now the send method will log the error with error_log() and fire
off a hook that can be used in WordPress for debugging purposes if desired.